remove vmobject.py and rename OpenGLVMobject -> VMobject
